pyproject.toml
[tool.poetry]
name = "resticrc"
version = "0.1.0"
description = "Configuration interface to a restic backup tool."
authors = ["Igor Ovsyannikov <kamish@outlook.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
click = "^7.0"
attrs = "^19.1"
pluggy = "^0.13.0"
PyYAML = "^5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.1"
pytest-mock = "^1.10"
pytest-cov = "^2.7"
black = "^19.10b0"

[tool.poetry.scripts]
resticrc = "resticrc.console:cli"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

